feat(rdr-120): P5.A.2 CatalogDB → CatalogStore shim (nexus-2t7o5)
Inverts the P5.A.1 wrapping. CatalogStore in nexus.db.t2.catalog is
now the authoritative owner of the .catalog.db SQLite handle;
nexus.catalog.catalog_db reduces to a thin re-export shim that
aliases CatalogDB to CatalogStore. Every existing
``from nexus.catalog.catalog_db import CatalogDB`` import continues
to resolve unchanged; the underlying class instance is the same
object that the T2Database.catalog property surfaces.

Moved verbatim from nexus.catalog.catalog_db to nexus.db.t2.catalog:
- _DOCUMENTS_FTS_TRIGGERS (FTS5 trigger DDL)
- _SCHEMA_SQL (full table layout: owners, documents, documents_fts,
  links, collections, _meta, document_chunks + indexes)
- The entire CatalogDB class body — sqlite3.connect, PRAGMAs,
  __init__ migration block (repo_root ADD COLUMN, nexus-7vuw
  owners PK-swap, source_mtime/alias_of/source_uri/bib_* ADDs,
  RDR-108 K1 document_chunks PK-swap, RDR-108 D2 collections
  auto-backfill), rebuild() / _rebuild_inner, next_document_number,
  search, descendants, execute, commit, transaction
  (@contextmanager), bulk_load_documents (@contextmanager), close.

RDR-108 invariants preserved by code-identity: the class body is
byte-for-byte the previous implementation under a new name.
Document.tumbler is doc identity; chunk natural ID =
sha256(chunk_text)[:32]; document_chunks manifest is authoritative.

§A8-exempt content writes (nexus_rdr/120-research-A9-catalog-extension):
- collections auto-backfill (structurally-bound-to-event-sourcing)
- owners PK-swap (structurally-bound-to-schema)
- document_chunks PK-swap (structurally-bound-to-schema)

Tumbler record imports (OwnerRecord/DocumentRecord/LinkRecord) are
TYPE_CHECKING-only in the substrate to avoid a runtime substrate →
consumer dependency.

Lint state:
- src/nexus/catalog/catalog_db.py — ZERO sqlite3.connect calls.
- src/nexus/catalog/synthesizer.py — still has 1 (P5.A.3 retires).
- catalog-allowlist count: 2 → 1 (P5.A.3 takes to 0).
- Total violations: 0 (unchanged).

Test updates:
- test_catalog_allowlist_count_metric: assert == 1.
- test_lint_reports_zero_violations_after_p4_cutover: assert
  catalog_allowlist_count == 1.

Validation:
- Full unit suite (excluding integration + stress): 7535 passed,
  34 skipped, 104 deselected.
- tests/db/, tests/daemon/, all catalog-related test files: 316
  passed.
- Stress harness (12 scenarios): green.

Out of scope (later sub-beads):
- P5.A.3 (nexus-nbsng): delete synthesizer.py:792 + lint = 0.
- P5.A.4 (nexus-uaq23): catalog test suite + indexer dogfood.
