+++ src/nexus/catalog/catalog_db.py
@@ -1,895 +1,24 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 # Copyright (c) 2026 Hal Hildebrand. All rights reserved.
+"""Legacy ``nexus.catalog.catalog_db`` import surface.
 
+RDR-120 P5.A.2 (nexus-2t7o5): the catalog SQLite layer moved into T2
+as the eighth domain store. ``CatalogDB`` is now an alias for
+:class:`nexus.db.t2.catalog.CatalogStore`; this module exists only as
+a re-export shim so existing ``from nexus.catalog.catalog_db import
+CatalogDB`` imports continue to resolve unchanged.
+
+No ``sqlite3.connect`` lives here — see :mod:`nexus.db.t2.catalog`
+for the authoritative implementation. P5.A.3 will remove the
+remaining direct catalog/ sqlite call sites and drop the catalog-
+allowlist count to zero.
+"""
 from __future__ import annotations
 
-import json
-import sqlite3
-import threading
-from contextlib import contextmanager
-from pathlib import Path
-
-import structlog
-
-from nexus.catalog.tumbler import DocumentRecord, LinkRecord, OwnerRecord
-from nexus.db.t2 import _sanitize_fts5
-
-_log = structlog.get_logger()
-
-# FTS5 trigger DDL extracted as standalone strings so the bulk-load
-# fence (``CatalogDB.bulk_load_documents``) can drop and recreate them
-# inside a transaction without re-running the entire schema script. Keep
-# in sync with the inline copies in ``_SCHEMA_SQL`` below — the inline
-# ones run on initial schema creation, these run on the rebuild path.
-_DOCUMENTS_FTS_TRIGGERS: tuple[str, ...] = (
-    """\
-CREATE TRIGGER IF NOT EXISTS documents_ai AFTER INSERT ON documents BEGIN
-    INSERT INTO documents_fts(rowid, title, author, corpus, file_path)
-        VALUES (new.rowid, new.title, new.author, new.corpus, new.file_path);
-END
-""",
-    """\
-CREATE TRIGGER IF NOT EXISTS documents_ad AFTER DELETE ON documents BEGIN
-    INSERT INTO documents_fts(documents_fts, rowid, title, author, corpus, file_path)
-        VALUES ('delete', old.rowid, old.title, old.author, old.corpus, old.file_path);
-END
-""",
-    """\
-CREATE TRIGGER IF NOT EXISTS documents_au AFTER UPDATE ON documents BEGIN
-    INSERT INTO documents_fts(documents_fts, rowid, title, author, corpus, file_path)
-        VALUES ('delete', old.rowid, old.title, old.author, old.corpus, old.file_path);
-    INSERT INTO documents_fts(rowid, title, author, corpus, file_path)
-        VALUES (new.rowid, new.title, new.author, new.corpus, new.file_path);
-END
-""",
+from nexus.db.t2.catalog import (  # noqa: F401 — re-exports
+    CatalogStore as CatalogDB,
+    _DOCUMENTS_FTS_TRIGGERS,
+    _SCHEMA_SQL,
 )
 
-_SCHEMA_SQL = """\
-PRAGMA journal_mode=WAL;
-
-CREATE TABLE IF NOT EXISTS owners (
-    tumbler_prefix TEXT PRIMARY KEY,
-    name TEXT NOT NULL,
-    owner_type TEXT NOT NULL,
-    repo_hash TEXT,
-    description TEXT,
-    repo_root TEXT DEFAULT '',
-    -- nexus-7vuw: name UNIQUE was a too-strict invariant. A repo and a
-    -- curator are different namespaces, so a repo named "nexus" should
-    -- coexist with a curator named "nexus" (e.g. ``nx index pdf
-    -- --corpus nexus`` after ``nx index repo .``). Pre-fix, the second
-    -- INSERT OR REPLACE silently obliterated the first row via the
-    -- name UNIQUE conflict, leaving owner_for_repo(repo_hash) returning
-    -- None and the indexer falling through to path-derived collection
-    -- naming. Composite UNIQUE keeps name-collision detection where it
-    -- belongs (within an owner_type).
-    UNIQUE(name, owner_type)
-);
-
-CREATE TABLE IF NOT EXISTS documents (
-    tumbler TEXT PRIMARY KEY,
-    title TEXT NOT NULL,
-    author TEXT,
-    year INTEGER,
-    content_type TEXT,
-    file_path TEXT,
-    corpus TEXT,
-    physical_collection TEXT,
-    chunk_count INTEGER,
-    head_hash TEXT,
-    indexed_at TEXT,
-    metadata JSON,
-    source_mtime REAL NOT NULL DEFAULT 0,
-    -- nexus-s8yz: permanent tumbler aliasing. When a document is
-    -- consolidated into a canonical owner (dedupe-owners, nexus-tmbh),
-    -- its row is kept and alias_of is set to the canonical tumbler.
-    -- External references (plan templates, prose citations, links
-    -- written by other systems) continue to resolve via alias_of —
-    -- that is the stability promise tumblers were chosen for.
-    -- '' (empty) means "this is the canonical document".
-    alias_of TEXT NOT NULL DEFAULT '',
-    -- RDR-096 P2.1: persistent URI identity. ``''`` (empty) on
-    -- legacy rows; populated for new registers after P2.1 ships.
-    -- Backfill derives URIs from ``file_path + physical_collection``.
-    source_uri TEXT NOT NULL DEFAULT '',
-    -- RDR-101 Phase 1 PR D (nexus-knn3): bibliographic enrichment
-    -- columns from the bib disposition deliverable
-    -- (docs/rdr/post-mortem/rdr-101-bib-disposition.md, Option A).
-    -- The bib_* fields move OFF T3 chunk metadata and live exactly once
-    -- on the Document projection. Phase 1 ships the empty columns;
-    -- Phase 3 wires DocumentEnriched v: 1 events to populate them
-    -- through the projector. The two indexed ID columns are the
-    -- "this title was enriched on backend X" cardinality marker that
-    -- nx enrich bib's skip query will read against (Phase 4); the
-    -- partial indexes (created below) make that query a sub-millisecond
-    -- presence test instead of a 300-row Chroma pagination.
-    bib_year INTEGER NOT NULL DEFAULT 0,
-    bib_authors TEXT NOT NULL DEFAULT '',
-    bib_venue TEXT NOT NULL DEFAULT '',
-    bib_citation_count INTEGER NOT NULL DEFAULT 0,
-    bib_semantic_scholar_id TEXT NOT NULL DEFAULT '',
-    bib_openalex_id TEXT NOT NULL DEFAULT '',
-    bib_doi TEXT NOT NULL DEFAULT '',
-    bib_enriched_at TEXT NOT NULL DEFAULT ''
-);
-
-CREATE VIRTUAL TABLE IF NOT EXISTS documents_fts USING fts5(
-    title, author, corpus, file_path,
-    content=documents, content_rowid=rowid
-);
-
-CREATE TRIGGER IF NOT EXISTS documents_ai AFTER INSERT ON documents BEGIN
-    INSERT INTO documents_fts(rowid, title, author, corpus, file_path)
-        VALUES (new.rowid, new.title, new.author, new.corpus, new.file_path);
-END;
-
-CREATE TRIGGER IF NOT EXISTS documents_ad AFTER DELETE ON documents BEGIN
-    INSERT INTO documents_fts(documents_fts, rowid, title, author, corpus, file_path)
-        VALUES ('delete', old.rowid, old.title, old.author, old.corpus, old.file_path);
-END;
-
-CREATE TRIGGER IF NOT EXISTS documents_au AFTER UPDATE ON documents BEGIN
-    INSERT INTO documents_fts(documents_fts, rowid, title, author, corpus, file_path)
-        VALUES ('delete', old.rowid, old.title, old.author, old.corpus, old.file_path);
-    INSERT INTO documents_fts(rowid, title, author, corpus, file_path)
-        VALUES (new.rowid, new.title, new.author, new.corpus, new.file_path);
-END;
-
-CREATE TABLE IF NOT EXISTS links (
-    id INTEGER PRIMARY KEY,
-    from_tumbler TEXT NOT NULL,
-    to_tumbler TEXT NOT NULL,
-    link_type TEXT NOT NULL,
-    from_span TEXT,
-    to_span TEXT,
-    created_by TEXT NOT NULL,
-    created_at TEXT,
-    metadata JSON
-);
-
-CREATE INDEX IF NOT EXISTS idx_links_from ON links(from_tumbler);
-CREATE INDEX IF NOT EXISTS idx_links_to ON links(to_tumbler);
-CREATE INDEX IF NOT EXISTS idx_links_type ON links(link_type);
-CREATE INDEX IF NOT EXISTS idx_links_created_by ON links(created_by);
-CREATE INDEX IF NOT EXISTS idx_links_from_type ON links(from_tumbler, link_type);
-CREATE INDEX IF NOT EXISTS idx_links_to_type ON links(to_tumbler, link_type);
-
-CREATE UNIQUE INDEX IF NOT EXISTS idx_links_unique
-    ON links(from_tumbler, to_tumbler, link_type);
-
-CREATE INDEX IF NOT EXISTS idx_links_created_by_type
-    ON links(created_by, link_type);
-
-CREATE INDEX IF NOT EXISTS idx_documents_tumbler
-    ON documents(tumbler);
-
--- RDR-101 Phase 6 (nexus-o6aa.14): first-class Collections projection.
--- One row per ChromaDB collection name. Materialized from
--- CollectionCreated events; legacy_grandfathered is projection-derived
--- from corpus.is_conformant_collection_name (no event-payload extension
--- required, v: 0 stays stable). Read paths consult this table to
--- distinguish post-Phase-6 canonical names from grandfathered legacy
--- names; write paths consult it to short-circuit re-registration.
-CREATE TABLE IF NOT EXISTS collections (
-    name TEXT PRIMARY KEY,
-    content_type TEXT NOT NULL DEFAULT '',
-    owner_id TEXT NOT NULL DEFAULT '',
-    embedding_model TEXT NOT NULL DEFAULT '',
-    model_version TEXT NOT NULL DEFAULT '',
-    display_name TEXT NOT NULL DEFAULT '',
-    -- 1 = name does NOT match is_conformant_collection_name; the row
-    -- exists only because the collection predates RDR-101 Phase 6 or
-    -- was manually registered by the operator. Read paths accept it.
-    legacy_grandfathered INTEGER NOT NULL DEFAULT 0,
-    superseded_by TEXT NOT NULL DEFAULT '',
-    superseded_at TEXT NOT NULL DEFAULT '',
-    created_at TEXT NOT NULL DEFAULT ''
-);
-
-CREATE INDEX IF NOT EXISTS idx_collections_legacy
-    ON collections(legacy_grandfathered);
-CREATE INDEX IF NOT EXISTS idx_collections_owner
-    ON collections(owner_id);
-
--- nexus-wehp: cross-process consistency-marker table. Stores the
--- highest canonical-source mtime that was successfully projected into
--- this SQLite cache. Catalog._ensure_consistent reads it on
--- construction to skip the DELETE+replay rebuild when the projection
--- is already up to date, eliminating the 'database is locked'
--- contention that surfaced when CLI write-side verbs raced an
--- nx-mcp-held connection in v4.23.0. A fresh SQLite cache has no
--- row, returns 0.0, and the rebuild fires (the e2e test invariant
--- 'fresh cache against existing catalog dir sees the data').
-CREATE TABLE IF NOT EXISTS _meta (
-    key TEXT PRIMARY KEY,
-    value TEXT
-);
-
--- RDR-103 Phase 2: ``Catalog.collection_for`` resolves a
--- ``(content_type, owner_id, embedding_model)`` triple to the
--- highest-versioned conformant collection. Without this index the
--- lookup is a full scan over the projection.
-CREATE INDEX IF NOT EXISTS idx_collections_tuple
-    ON collections(content_type, owner_id, embedding_model);
-
--- RDR-101 Phase 1 PR D (nexus-knn3) partial indexes on bib backend IDs
--- live in the post-migration block in __init__: the legacy-DB upgrade
--- path has to ALTER TABLE the bib columns into existence before the
--- partial-index CREATE can reference them.
-
--- RDR-108 D2 (nexus-mydi): document_chunks manifest. The catalog is
--- the authoritative source of truth for doc->chunk ordering (the
--- "tree" layer of the git/IPFS-style blob+tree split). T3 chunks are
--- content-addressed blobs keyed on chunk_text_hash[:32]; this table
--- records the ordered (doc_id, position) -> chash references that
--- compose each Document. The same chash can appear at multiple
--- (doc_id, position) rows: the manifest preserves position; T3
--- stores content once. Optional positional columns (line_start /
--- line_end / char_start / char_end) carry display-friendly span
--- coordinates so retrieval doesn't have to re-derive them from the
--- source file. chunk_index is the chunker-assigned ordinal at index
--- time, retained for reference; position is the canonical ordering
--- key from this RDR onward.
-CREATE TABLE IF NOT EXISTS document_chunks (
-    doc_id      TEXT NOT NULL REFERENCES documents(tumbler) ON DELETE CASCADE,
-    position    INTEGER NOT NULL,
-    chash       TEXT NOT NULL,
-    chunk_index INTEGER,
-    line_start  INTEGER,
-    line_end    INTEGER,
-    char_start  INTEGER,
-    char_end    INTEGER,
-    PRIMARY KEY (doc_id, position)
-);
-
-CREATE INDEX IF NOT EXISTS idx_document_chunks_chash
-    ON document_chunks(chash);
-"""
-
-
-class CatalogDB:
-    """SQLite query cache for the JSONL-backed catalog."""
-
-    def __init__(self, db_path: Path) -> None:
-        self._path = db_path
-        self._conn = sqlite3.connect(str(db_path), check_same_thread=False)
-        # Reentrant: ``execute()`` callers inside a ``transaction()``
-        # context (RDR-101 round-3) re-acquire the lock from the same
-        # thread. With a plain ``Lock()`` the nested ``with self._lock:``
-        # in ``execute`` would deadlock. ``RLock`` is a strict superset
-        # of ``Lock`` for cross-thread mutual exclusion.
-        self._lock = threading.RLock()
-        # Storage review I-2: match the T2 domain-store concurrency defaults
-        # (5 s busy_timeout + WAL) so cross-process writers don't immediately
-        # raise ``OperationalError: database is locked`` when the CLI races
-        # an indexing writer.
-        self._conn.execute("PRAGMA busy_timeout=5000")
-        self._conn.execute("PRAGMA journal_mode=WAL")
-        # RDR-108 D2 + D5 (nexus-mydi): enable foreign-key enforcement
-        # so the document_chunks manifest cannot reference a deleted
-        # Document, and future cascade-on-rename machinery (D5
-        # chash_index FK) works as advertised. SQLite enforces FK
-        # constraints only when this PRAGMA is ON; declared REFERENCES
-        # clauses in the schema are otherwise advisory.
-        self._conn.execute("PRAGMA foreign_keys=ON")
-        # Issue #437: cap WAL growth under long-lived MCP-server reader
-        # connections. SQLite's auto-checkpoint can run only as PASSIVE
-        # while readers hold pre-checkpoint snapshots; PASSIVE folds
-        # frames into the main DB but cannot truncate the WAL file.
-        # Without a journal_size_limit the WAL grows unbounded over a
-        # multi-hour session (12 MB observed on the reporter's install
-        # before manual ``PRAGMA wal_checkpoint(TRUNCATE)``). 64 MiB
-        # caps the steady-state size after each successful checkpoint;
-        # SQLite reuses the space rather than growing the file.
-        self._conn.execute("PRAGMA journal_size_limit=67108864")
-        self._conn.executescript(_SCHEMA_SQL)
-        # Migration: add repo_root column if missing (pre-RDR-060 databases)
-        try:
-            self._conn.execute("SELECT repo_root FROM owners LIMIT 0")
-        except sqlite3.OperationalError:
-            with self._conn:
-                self._conn.execute("ALTER TABLE owners ADD COLUMN repo_root TEXT DEFAULT ''")
-
-        # nexus-7vuw: drop the legacy single-column UNIQUE(name) index
-        # on owners (replaced with composite UNIQUE(name, owner_type) in
-        # the schema above). Pre-fix DBs carry a sqlite_autoindex_owners_*
-        # index backing the old single-column constraint; SQLite cannot
-        # ALTER a constraint in place, so the migration rebuilds the
-        # owners table from existing rows. Idempotent: skipped when the
-        # auto-index is already absent.
-        legacy_unique = self._conn.execute(
-            "SELECT name FROM sqlite_master WHERE type='index' "
-            "AND tbl_name='owners' AND name LIKE 'sqlite_autoindex_owners_%' "
-            "AND sql IS NULL"
-        ).fetchall()
-        if legacy_unique:
-            # Detect: probe the auto-index. If it indexes only ``name``,
-            # we have the pre-fix shape and need to rebuild. The
-            # composite UNIQUE(name, owner_type) declared in the schema
-            # above creates an auto-index on TWO columns, so single-column
-            # auto-indexes can only mean the legacy constraint survived
-            # from an older DB.
-            needs_rebuild = False
-            for (idx_name,) in legacy_unique:
-                cols = self._conn.execute(
-                    f"PRAGMA index_info({idx_name!r})"
-                ).fetchall()
-                if len(cols) == 1 and cols[0][2] == "name":
-                    needs_rebuild = True
-                    break
-            if needs_rebuild:
-                with self._conn:
-                    self._conn.execute(
-                        "CREATE TABLE owners_nexus7vuw_new ("
-                        "    tumbler_prefix TEXT PRIMARY KEY, "
-                        "    name TEXT NOT NULL, "
-                        "    owner_type TEXT NOT NULL, "
-                        "    repo_hash TEXT, "
-                        "    description TEXT, "
-                        "    repo_root TEXT DEFAULT '', "
-                        "    UNIQUE(name, owner_type)"
-                        ")"
-                    )
-                    # INSERT OR IGNORE so any pre-existing colliding rows
-                    # (which the legacy UNIQUE name constraint would have
-                    # already prevented) are dropped silently rather than
-                    # tripping the new composite constraint.
-                    self._conn.execute(
-                        "INSERT OR IGNORE INTO owners_nexus7vuw_new "
-                        "(tumbler_prefix, name, owner_type, repo_hash, "
-                        " description, repo_root) "
-                        "SELECT tumbler_prefix, name, owner_type, repo_hash, "
-                        "       description, repo_root FROM owners"
-                    )
-                    self._conn.execute("DROP TABLE owners")
-                    self._conn.execute(
-                        "ALTER TABLE owners_nexus7vuw_new RENAME TO owners"
-                    )
-
-        # nexus-8luh: add source_mtime column to existing databases so
-        # RDR-087 Phase 3.4's stale_source_ratio has something to read
-        # from. Default 0 for pre-migration rows (meaning "unknown").
-        try:
-            self._conn.execute("SELECT source_mtime FROM documents LIMIT 0")
-        except sqlite3.OperationalError:
-            with self._conn:
-                self._conn.execute(
-                    "ALTER TABLE documents ADD COLUMN source_mtime REAL NOT NULL DEFAULT 0"
-                )
-
-        # nexus-s8yz: add alias_of column to existing databases. '' means
-        # the document is canonical (not an alias).
-        try:
-            self._conn.execute("SELECT alias_of FROM documents LIMIT 0")
-        except sqlite3.OperationalError:
-            with self._conn:
-                self._conn.execute(
-                    "ALTER TABLE documents ADD COLUMN alias_of TEXT NOT NULL DEFAULT ''"
-                )
-
-        # RDR-096 P2.1 (nexus-ocu9.3): add source_uri column to existing
-        # databases. ``''`` on pre-migration rows; new rows get
-        # populated at register time. Backfill happens lazily on
-        # rebuild from JSONL — DocumentRecord.source_uri carries the
-        # value through.
-        try:
-            self._conn.execute("SELECT source_uri FROM documents LIMIT 0")
-        except sqlite3.OperationalError:
-            with self._conn:
-                self._conn.execute(
-                    "ALTER TABLE documents ADD COLUMN source_uri TEXT NOT NULL DEFAULT ''"
-                )
-
-        # RDR-101 Phase 1 PR D (nexus-knn3): add bib_* columns to existing
-        # databases. The bib disposition deliverable
-        # (docs/rdr/post-mortem/rdr-101-bib-disposition.md, Option A)
-        # moves these fields off T3 chunk metadata and onto the Document
-        # projection. Phase 1 ships the columns empty; Phase 3 wires the
-        # projector to populate them from DocumentEnriched v: 1 events.
-        # Each ALTER probes for the column first; failure means it was
-        # added in a previous run (idempotent migration pattern matches
-        # the rest of this method).
-        for col_name, col_decl in (
-            ("bib_year",                "INTEGER NOT NULL DEFAULT 0"),
-            ("bib_authors",             "TEXT NOT NULL DEFAULT ''"),
-            ("bib_venue",               "TEXT NOT NULL DEFAULT ''"),
-            ("bib_citation_count",      "INTEGER NOT NULL DEFAULT 0"),
-            ("bib_semantic_scholar_id", "TEXT NOT NULL DEFAULT ''"),
-            ("bib_openalex_id",         "TEXT NOT NULL DEFAULT ''"),
-            ("bib_doi",                 "TEXT NOT NULL DEFAULT ''"),
-            ("bib_enriched_at",         "TEXT NOT NULL DEFAULT ''"),
-        ):
-            try:
-                self._conn.execute(f"SELECT {col_name} FROM documents LIMIT 0")
-            except sqlite3.OperationalError:
-                with self._conn:
-                    self._conn.execute(
-                        f"ALTER TABLE documents ADD COLUMN {col_name} {col_decl}"
-                    )
-
-        # Partial indexes on the two bib backend IDs. CREATE INDEX IF NOT
-        # EXISTS is safe on a fresh DB (where _SCHEMA_SQL already created
-        # them) and on existing DBs (where the columns just landed via
-        # ALTER above).
-        with self._conn:
-            self._conn.execute(
-                "CREATE INDEX IF NOT EXISTS idx_documents_bib_s2_id "
-                "ON documents(bib_semantic_scholar_id) "
-                "WHERE bib_semantic_scholar_id != ''"
-            )
-            self._conn.execute(
-                "CREATE INDEX IF NOT EXISTS idx_documents_bib_oa_id "
-                "ON documents(bib_openalex_id) "
-                "WHERE bib_openalex_id != ''"
-            )
-
-        # RDR-108 Phase 4 (nexus-dyxe): index documents.physical_collection
-        # so manifest-based GC's per-collection lookup
-        # (``chashes_for_collection`` joins document_chunks to documents
-        # filtered by physical_collection) does not full-scan documents.
-        # Same column is the filter key for ``list_by_collection`` and
-        # ``relocate_collection``. Probe-guarded so stripped-down legacy
-        # schemas without ``physical_collection`` (e.g. the alias_of
-        # migration test fixture) still open cleanly.
-        try:
-            self._conn.execute("SELECT physical_collection FROM documents LIMIT 0")
-        except sqlite3.OperationalError:
-            pass
-        else:
-            with self._conn:
-                self._conn.execute(
-                    "CREATE INDEX IF NOT EXISTS idx_documents_physical_collection "
-                    "ON documents(physical_collection)"
-                )
-
-        # RDR-108 K1 (nexus-lh8c): add ON DELETE CASCADE to document_chunks
-        # for existing databases that were created before this constraint was
-        # declared in the schema. SQLite cannot ALTER a FK constraint in place;
-        # the 12-step pattern recreates the table with the correct declaration.
-        # Idempotent: skipped if the schema already includes ON DELETE CASCADE.
-        _chunks_ddl_row = self._conn.execute(
-            "SELECT sql FROM sqlite_master WHERE type='table' AND name='document_chunks'"
-        ).fetchone()
-        if _chunks_ddl_row is not None and "ON DELETE CASCADE" not in _chunks_ddl_row[0]:
-            # Detect which columns exist — older schemas may be a strict subset.
-            _old_col_names = {
-                r[1] for r in self._conn.execute(
-                    "PRAGMA table_info(document_chunks)"
-                ).fetchall()
-            }
-            _all_cols = [
-                "doc_id", "position", "chash", "chunk_index",
-                "line_start", "line_end", "char_start", "char_end",
-            ]
-            _copy_cols = [c for c in _all_cols if c in _old_col_names]
-            _copy_sql = ", ".join(_copy_cols)
-            with self._conn:
-                self._conn.execute(
-                    "CREATE TABLE document_chunks_rdr108k1_new ("
-                    "    doc_id      TEXT NOT NULL"
-                    "                REFERENCES documents(tumbler) ON DELETE CASCADE,"
-                    "    position    INTEGER NOT NULL,"
-                    "    chash       TEXT NOT NULL,"
-                    "    chunk_index INTEGER,"
-                    "    line_start  INTEGER,"
-                    "    line_end    INTEGER,"
-                    "    char_start  INTEGER,"
-                    "    char_end    INTEGER,"
-                    "    PRIMARY KEY (doc_id, position)"
-                    ")"
-                )
-                self._conn.execute(
-                    f"INSERT INTO document_chunks_rdr108k1_new ({_copy_sql})"
-                    f" SELECT {_copy_sql} FROM document_chunks"
-                )
-                self._conn.execute("DROP TABLE document_chunks")
-                self._conn.execute(
-                    "ALTER TABLE document_chunks_rdr108k1_new RENAME TO document_chunks"
-                )
-                self._conn.execute(
-                    "CREATE INDEX IF NOT EXISTS idx_document_chunks_chash"
-                    " ON document_chunks(chash)"
-                )
-
-                # RDR-108 D2 (nexus-mydi): backfill collections rows for any
-        # documents.physical_collection value that has no matching
-        # collections.name row. Pre-RDR-108 catalogs accumulated docs
-        # whose physical_collection was a free-form string, never
-        # registered in the collections projection — surfaced as the
-        # 11.3x chash_index drift and 76% document_aspects orphan rate
-        # in the 2026-05-08 prod-shakeout. This backfill is structural
-        # prep for FK enforcement on documents.physical_collection
-        # (deferred to a follow-up migration that recreates the
-        # documents table per SQLite's 12-step ALTER pattern).
-        # Idempotent: the SELECT-then-INSERT pattern with a
-        # collections.name PK on the target side is a no-op when the
-        # row already exists.
-        # Probe-guarded: stripped-down legacy schemas (e.g. the
-        # alias_of migration test fixture) may lack the
-        # physical_collection column entirely; skip the backfill in
-        # that case rather than raise.
-        # nexus-572g K7: track inserted names so Catalog.__init__ can
-        # emit synthetic CollectionCreated events with
-        # legacy_grandfathered=True. Without backing events the rows
-        # vanish on Catalog.rebuild() (DELETE FROM collections + JSONL
-        # replay). Catalog reads _backfilled_collections immediately
-        # after constructing this CatalogDB.
-        self._backfilled_collections: set[str] = set()
-        try:
-            self._conn.execute(
-                "SELECT physical_collection FROM documents LIMIT 0"
-            )
-        except sqlite3.OperationalError:
-            pass
-        else:
-            # RDR-108 D2 (nexus-mydi → nexus-572g): identify candidate
-            # collection names before INSERT so we track exactly which
-            # names actually land (the INSERT below skips existing rows).
-            # The candidate set seeds _emit_backfilled_collection_events
-            # (catalog.py) so the event-sourced model stays canonical:
-            # synthetic CollectionCreated events are emitted post-rebuild
-            # so a JSONL replay materializes the same collections rows.
-            #
-            # SIG-7 (nexus-872w): created_at is a real ISO timestamp
-            # (parameter-bound, not f-string interpolated) so audit
-            # tools can distinguish rows that were backfilled from
-            # rows that were never written.
-            #
-            # O-5: INSERT OR IGNORE replaces the NOT IN subquery
-            # form (perf — collections.name is a PK so OR IGNORE is
-            # the natural shape).
-            candidates = {
-                row[0]
-                for row in self._conn.execute(
-                    "SELECT DISTINCT physical_collection FROM documents "
-                    "WHERE physical_collection IS NOT NULL "
-                    "  AND physical_collection != '' "
-                    "  AND physical_collection NOT IN "
-                    "      (SELECT name FROM collections)"
-                ).fetchall()
-            }
-            if candidates:
-                # nexus-33xm: created_at is INTENTIONALLY left empty
-                # here. The companion ``_emit_backfilled_collection_events``
-                # (catalog.py) emits CollectionCreated events with
-                # ``created_at=""`` for these same names; the projector
-                # handler does ``INSERT OR REPLACE ... COALESCE((SELECT
-                # created_at FROM collections WHERE name = ?), ?)``. If
-                # this auto-bootstrap stamps ``NOW()``, the COALESCE
-                # preserves the synthetic stamp on event apply but a
-                # fresh ``--replay-equality`` replay (which starts from
-                # an empty table) takes the ``""`` from the event
-                # payload, producing a permanent drift between live
-                # and projected. Empty here keeps the two paths
-                # bit-equal.
-                #
-                # SIG-7 (nexus-872w) audit-distinction goal: the prior
-                # NOW() stamp was meant to let audit tools distinguish
-                # auto-bootstrapped rows from event-derived rows. That
-                # distinction now lives in the synthetic event itself
-                # (``payload.created_at == ""`` marks a backfilled row);
-                # the SQLite row no longer needs to carry it.
-                with self._conn:
-                    self._conn.execute(
-                        "INSERT OR IGNORE INTO collections "
-                        "(name, content_type, owner_id, embedding_model, "
-                        " model_version, display_name, legacy_grandfathered, "
-                        " superseded_by, superseded_at, created_at) "
-                        "SELECT DISTINCT physical_collection, '', '', '', '', '', "
-                        "  1, '', '', '' "
-                        "FROM documents "
-                        "WHERE physical_collection IS NOT NULL "
-                        "  AND physical_collection != ''",
-                    )
-                self._backfilled_collections = candidates
-
-    def rebuild(
-        self,
-        owners: dict[str, OwnerRecord],
-        documents: dict[str, DocumentRecord],
-        links: list[LinkRecord],
-        *,
-        consistency_mtime: float | None = None,
-    ) -> None:
-        """Truncate all tables and reload from JSONL-derived dicts.
-
-        Uses the FTS5 bulk-load fence (drop triggers + INSERT-rebuild)
-        so the per-row ``documents_ai`` trigger does NOT queue every
-        column for every replayed document — that pattern stalls COMMIT
-        for tens of minutes on a catalog with hundreds of thousands of
-        rows. With the fence, FTS5 segments are built once at the end
-        in source order. See ``bulk_load_documents`` docstring.
-
-        ``consistency_mtime`` (RDR-104 critic Critical #2 fix): when
-        supplied, the consistency-marker write happens INSIDE the same
-        ``with self._lock, self._conn:`` block as the projection writes
-        — atomic. Pre-fix the marker was written by an independent
-        ``commit()`` after rebuild returned, which left a refactoring
-        hazard (any future code that put the marker write before the
-        projection commit would silently corrupt the projection by
-        skipping events on the next run). Putting both writes in the
-        same transaction makes the atomicity invariant trivially
-        true regardless of caller ordering.
-        """
-        # RDR-108 Phase 3 (nexus-bdag): ``document_chunks`` is FK-bound
-        # to ``documents`` with ON DELETE CASCADE. The DELETE FROM
-        # documents below would cascade-wipe the manifest, but the
-        # projector does not re-emit ``ChunkIndexed`` rows during
-        # legacy replay (the manifest is populated by the post-store
-        # batch hook, not by the JSONL log yet). Disable FK enforcement
-        # around the DELETE+reload so the cascade doesn't fire; INSERTs
-        # restore valid references and we re-enable FK afterwards.
-        # PRAGMA foreign_keys is a no-op within a transaction, so it
-        # must run BEFORE the ``with self._conn:`` block opens its
-        # transaction.
-        self._conn.execute("PRAGMA foreign_keys=OFF")
-        try:
-            self._rebuild_inner(owners, documents, links, consistency_mtime=consistency_mtime)
-            # nexus-lrhg #3: with FK enforcement disabled the DELETE FROM
-            # documents above did NOT cascade-delete the
-            # ``document_chunks`` manifest. Any rows whose ``doc_id``
-            # references a tombstoned document (DocumentDeleted in the
-            # legacy log that this rebuild replays) survive the wipe and
-            # become orphans: they reference a doc_id that the new
-            # documents INSERT loop did not restore. PRAGMA foreign_keys
-            # =ON below only enforces new writes, not existing rows, so
-            # the orphan rows persist silently forever. Wipe them
-            # explicitly before re-enabling FK so the post-rebuild state
-            # is FK-clean.
-            orphan_count = self._conn.execute(
-                "SELECT COUNT(*) FROM document_chunks "
-                "WHERE doc_id NOT IN (SELECT tumbler FROM documents)"
-            ).fetchone()[0]
-            if orphan_count:
-                self._conn.execute(
-                    "DELETE FROM document_chunks "
-                    "WHERE doc_id NOT IN (SELECT tumbler FROM documents)"
-                )
-                self._conn.commit()
-                _log.info(
-                    "catalog_db_rebuild_orphan_chunks_purged",
-                    count=orphan_count,
-                )
-        finally:
-            self._conn.execute("PRAGMA foreign_keys=ON")
-
-        _log.debug("catalog_db.rebuild", owners=len(owners), documents=len(documents), links=len(links))
-
-    def _rebuild_inner(
-        self,
-        owners: dict[str, OwnerRecord],
-        documents: dict[str, DocumentRecord],
-        links: list[LinkRecord],
-        *,
-        consistency_mtime: float | None,
-    ) -> None:
-        """Inner rebuild routine called by :meth:`rebuild` with FK
-        enforcement disabled around it (RDR-108 Phase 3)."""
-        with self._lock, self._conn, self.bulk_load_documents():
-            # Delete from base tables. Triggers are dropped for the
-            # duration of the bulk_load_documents fence; FTS5 is
-            # rebuilt in one pass at fence-exit.
-            self._conn.execute("DELETE FROM links")
-            self._conn.execute("DELETE FROM documents")
-            self._conn.execute("DELETE FROM owners")
-            # RDR-104 Step 0 (Critical #1): clear ``collections`` too.
-            # Legacy rebuild has no ``collections`` dict (collections
-            # are event-sourced exclusively), so this DELETE leaves
-            # the table empty for the event-sourced replay that
-            # always follows when ``events.jsonl`` exists. Pre-fix
-            # the absence of this DELETE combined with
-            # ``_v0_collection_created``'s ``INSERT OR REPLACE`` plus
-            # COALESCE-preservation pattern silently inherited stale
-            # ``superseded_by``/``superseded_at``/``created_at`` from
-            # the prior run.
-            self._conn.execute("DELETE FROM collections")
-
-            for prefix, o in owners.items():
-                self._conn.execute(
-                    "INSERT OR REPLACE INTO owners (tumbler_prefix, name, owner_type, repo_hash, description, repo_root) "
-                    "VALUES (?, ?, ?, ?, ?, ?)",
-                    (prefix, o.name, o.owner_type, o.repo_hash, o.description, o.repo_root),
-                )
-
-            for tumbler, d in documents.items():
-                self._conn.execute(
-                    "INSERT INTO documents "
-                    "(tumbler, title, author, year, content_type, file_path, "
-                    "corpus, physical_collection, chunk_count, head_hash, indexed_at, "
-                    "metadata, source_mtime, alias_of, source_uri) "
-                    "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                    (
-                        tumbler,
-                        d.title,
-                        d.author,
-                        d.year,
-                        d.content_type,
-                        d.file_path,
-                        d.corpus,
-                        d.physical_collection,
-                        d.chunk_count,
-                        d.head_hash,
-                        d.indexed_at,
-                        json.dumps(d.meta),
-                        d.source_mtime,
-                        d.alias_of,
-                        d.source_uri,
-                    ),
-                )
-
-            for lnk in links:
-                self._conn.execute(
-                    "INSERT OR IGNORE INTO links "
-                    "(from_tumbler, to_tumbler, link_type, from_span, to_span, "
-                    "created_by, created_at, metadata) "
-                    "VALUES (?, ?, ?, ?, ?, ?, ?, ?)",
-                    (
-                        lnk.from_t,
-                        lnk.to_t,
-                        lnk.link_type,
-                        lnk.from_span,
-                        lnk.to_span,
-                        lnk.created_by,
-                        lnk.created_at,
-                        json.dumps(lnk.meta),
-                    ),
-                )
-
-            # RDR-104 critic Critical #2 fix: consistency marker write
-            # is part of the same transaction as the projection writes.
-            # On a mid-rebuild crash, the projection rolls back AND the
-            # marker stays at its pre-rebuild value — next run sees the
-            # stale marker and re-rebuilds correctly. Pre-fix this lived
-            # outside the transaction in `_write_consistency_marker`'s
-            # own commit().
-            if consistency_mtime is not None:
-                self._conn.execute(
-                    "INSERT OR REPLACE INTO _meta (key, value) VALUES (?, ?)",
-                    ("last_consistency_mtime", f"{consistency_mtime}"),
-                )
-
-    def next_document_number(self, owner_prefix: str) -> int:
-        """Max document number for owner + 1.
-
-        Uses dot-count matching to avoid lexicographic ordering bugs
-        (e.g., '1.10' < '1.9' in string comparison).
-        """
-        depth = len(owner_prefix.split("."))
-        with self._lock:
-            row = self._conn.execute(
-                "SELECT MAX(CAST(substr(tumbler, length(?) + 2) AS INTEGER)) "
-                "FROM documents WHERE tumbler LIKE ? "
-                "AND (length(tumbler) - length(replace(tumbler, '.', ''))) = ?",
-                (owner_prefix, owner_prefix + ".%", depth),
-            ).fetchone()
-            return (row[0] or 0) + 1
-
-    def search(self, query: str, *, content_type: str | None = None) -> list[dict]:
-        """FTS5 MATCH over title, author, corpus, file_path."""
-        safe_q = _sanitize_fts5(query)
-        if not safe_q.strip():
-            return []
-
-        with self._lock:
-            sql = (
-                "SELECT d.tumbler, d.title, d.author, d.year, d.content_type, "
-                "d.file_path, d.corpus, d.physical_collection, d.chunk_count, "
-                "d.head_hash, d.indexed_at, d.metadata, d.source_mtime "
-                "FROM documents d "
-                "JOIN documents_fts f ON d.rowid = f.rowid "
-                "WHERE documents_fts MATCH ?"
-            )
-            params: list[str] = [safe_q]
-
-            if content_type:
-                sql += " AND d.content_type = ?"
-                params.append(content_type)
-
-            rows = self._conn.execute(sql, params).fetchall()
-            columns = ["tumbler", "title", "author", "year", "content_type",
-                        "file_path", "corpus", "physical_collection", "chunk_count",
-                        "head_hash", "indexed_at", "metadata", "source_mtime"]
-            return [dict(zip(columns, row)) for row in rows]
-
-    def descendants(self, prefix: str) -> list[dict]:
-        """All documents whose tumbler starts with prefix (any depth).
-
-        Uses LIKE 'prefix.%' so prefix itself is excluded — only strict
-        descendants are returned.
-        """
-        with self._lock:
-            rows = self._conn.execute(
-                "SELECT tumbler, title, author, year, content_type, "
-                "file_path, corpus, physical_collection, chunk_count, "
-                "head_hash, indexed_at, metadata, source_mtime "
-                "FROM documents WHERE tumbler LIKE ?",
-                (prefix + ".%",),
-            ).fetchall()
-        columns = [
-            "tumbler", "title", "author", "year", "content_type",
-            "file_path", "corpus", "physical_collection", "chunk_count",
-            "head_hash", "indexed_at", "metadata", "source_mtime",
-        ]
-        return [dict(zip(columns, row)) for row in rows]
-
-    def execute(self, sql: str, params: tuple | list = ()) -> sqlite3.Cursor:
-        """Thread-safe execute wrapper. Acquires _lock before executing."""
-        with self._lock:
-            return self._conn.execute(sql, params)
-
-    def commit(self) -> None:
-        """Thread-safe commit wrapper."""
-        with self._lock:
-            self._conn.commit()
-
-    @contextmanager
-    def transaction(self):
-        """Atomic transaction context (RDR-101 round-3, atomicity fix).
-
-        Wraps the connection in ``with self._lock, self._conn:`` so a
-        sequence of ``execute()`` calls runs as one transaction —
-        commits on success, rolls back on any exception. ``_lock`` is
-        an ``RLock`` so the inner ``execute()`` calls re-acquire the
-        same thread's lock without deadlock.
-
-        Used by ``Catalog._ensure_consistent`` to make the
-        DELETE+replay rebuild atomic; pre-fix the three DELETEs
-        autocommitted before ``Projector.apply_all`` ran, so any
-        exception mid-replay (a malformed event, the v: 1 raise) left
-        SQLite empty and unrecoverable until the next mtime tick.
-        """
-        with self._lock, self._conn:
-            yield self._conn
-
-    @contextmanager
-    def bulk_load_documents(self):
-        """FTS5 bulk-load fence around mass document writes.
-
-        Drops the ``documents_ai`` / ``documents_au`` / ``documents_ad``
-        FTS5 triggers, yields to the caller (which performs the writes),
-        then recreates the triggers and runs FTS5's ``rebuild`` command
-        to materialize the index in one pass from the final document
-        rows.
-
-        Why this exists: the projection-rebuild path in
-        ``Catalog._ensure_consistent`` deletes every document and replays
-        the entire event log inside one transaction. With per-row FTS5
-        triggers active, each replayed INSERT queues every term/column
-        in an in-memory hash. SQLite cannot incrementally merge that
-        hash into on-disk segments mid-transaction; the merge happens at
-        COMMIT, in one shot, and walks every queued entry. On a project
-        with hundreds of thousands of events the COMMIT alone takes
-        15-20 minutes of CPU on ``fts5IndexCrisismerge`` /
-        ``fts5HashEntrySort``, with no user-visible signal that anything
-        is happening.
-
-        FTS5's documented bulk-load idiom — ``INSERT INTO fts(fts) VALUES
-        ('rebuild')`` — is dramatically faster because it builds segments
-        directly from the content table in source order, skipping the
-        per-row hash queue entirely.
-
-        MUST be called inside a ``transaction()`` block so the schema
-        changes (DROP/CREATE TRIGGER) and the rebuild are part of the
-        same atomic unit as the DELETE+replay. If the caller's transaction
-        rolls back, the triggers come back too.
-        """
-        with self._lock:
-            for trig in ("documents_ai", "documents_au", "documents_ad"):
-                self._conn.execute(f"DROP TRIGGER IF EXISTS {trig}")
-        try:
-            yield
-        finally:
-            with self._lock:
-                for sql in _DOCUMENTS_FTS_TRIGGERS:
-                    self._conn.execute(sql)
-                # Bulk-rebuild FTS5 from the final state of `documents`.
-                # Cheap when the table is small; orders of magnitude
-                # faster than the per-row trigger path on rebuilds with
-                # >1k documents.
-                self._conn.execute(
-                    "INSERT INTO documents_fts(documents_fts) VALUES ('rebuild')"
-                )
-
-    def close(self) -> None:
-        with self._lock:
-            self._conn.close()
+__all__ = ("CatalogDB", "_DOCUMENTS_FTS_TRIGGERS", "_SCHEMA_SQL")
